refactor(controlplane): remove dynamic warm-capacity targeting + demand signal
First stage of tearing out the warm worker pool. Workers in the remote/k8s
backend are now only ever hot-active or hot-idle, spawned on demand and reused
by exact profile match — there is no warm pool to size, so the configstore-driven
dynamic warm-capacity target computation and its demand (warm-miss) signal are
dead.

Removed:
- warm_capacity_targets.go / warm_capacity_targets_k8s.go (+ tests) — the dynamic
  target computation (base + miss-driven extras, ceilings, headroom).
- multitenant_warm_capacity_test.go and the computeEffectiveWarmCapacityTarget*,
  warmCapacityGlobalCapBlocksDemand, reconcileWarmCapacityImageTargets helpers.
- OrgRouter.reconcileWarmCapacity / computeBaseWarmCapacityTargets /
  computePerImageWarmTargets and their callers on snapshot apply.
- K8sWorkerPool.recordWarmCapacityMiss + the per-miss demand recording in
  ReserveSharedWorker.
- The janitor's warm-capacity-miss-bucket pruner and reconcileWarmCapacity hook,
  replaced by a slim observeWorkerLifecycle hook that still refreshes the
  per-image worker lifecycle gauges (pure observability).

Renamed warm_capacity_metrics.go → pool_metrics.go (keeps only the worker
lifecycle gauge; the warm-target/headroom/miss gauges are gone) and
warm_pool_metrics_test.go → pool_metrics_test.go.

The warm-pool *spawn* machinery (SpawnMinWorkers, SetWarmCapacityTarget, neutral
slots, replenish) and colocate/tiers come out in following stages; the configstore
warm methods and warm config fields are now unused and removed there too. Build +
vet + controlplane unit tests green on both tag sets; process backend untouched.
