+++ controlplane/org_router_test.go
@@ -132,37 +132,6 @@ func TestOrgRouterDestroyOrgStackDrainsSessionsBeforePoolShutdownAndReleasesSess
 	}
 }
 
-func TestOrgRouterReconcileWarmCapacityUsesExplicitSharedWarmTarget(t *testing.T) {
-	sharedPool, _ := newTestK8sPool(t, 10)
-	sharedPool.spawnWarmWorkerFunc = func(ctx context.Context, id int) error {
-		sharedPool.mu.Lock()
-		defer sharedPool.mu.Unlock()
-		sharedPool.workers[id] = &ManagedWorker{ID: id, done: make(chan struct{})}
-		return nil
-	}
-	tr := &OrgRouter{
-		sharedPool: sharedPool,
-		globalCfg: ControlPlaneConfig{
-			K8s: K8sConfig{
-				SharedWarmTarget: 4,
-			},
-		},
-	}
-
-	snap := &configstore.Snapshot{
-		Orgs: map[string]*configstore.OrgConfig{
-			"analytics": {Name: "analytics"},
-			"billing":   {Name: "billing"},
-		},
-	}
-
-	tr.reconcileWarmCapacity(snap)
-
-	if got := sharedPool.minWorkers; got != 4 {
-		t.Fatalf("expected shared warm target 4, got %d", got)
-	}
-}
-
 func TestOrgRouterAdapterSetWarmCapacityTargetZeroClearsAllWarmTargets(t *testing.T) {
 	sharedPool, _ := newTestK8sPool(t, 10)
 	sharedPool.SetWarmCapacityTarget(4)
@@ -457,157 +426,3 @@ func TestWorkerImageForOrg(t *testing.T) {
 	}
 }
 
-func TestOrgRouterReconcileWarmCapacityFloorsOnePerActiveImage(t *testing.T) {
-	sharedPool, _ := newTestK8sPool(t, 10)
-
-	tr := &OrgRouter{
-		sharedPool: sharedPool,
-		baseCfg: K8sWorkerPoolConfig{
-			WorkerImage: "posthog/duckgres:default",
-		},
-		globalCfg: ControlPlaneConfig{K8s: K8sConfig{SharedWarmTarget: 1}},
-		orgs: map[string]*OrgStack{
-			"analytics": {Config: &configstore.OrgConfig{Name: "analytics"}},
-			"billing":   {Config: &configstore.OrgConfig{Name: "billing"}},
-			// orphans-without-stack — should NOT contribute to per-image floor
-		},
-	}
-
-	snap := &configstore.Snapshot{
-		Orgs: map[string]*configstore.OrgConfig{
-			"analytics": {
-				Name: "analytics",
-				Warehouse: &configstore.ManagedWarehouseConfig{
-					Image: "posthog/duckgres:v1.5.1",
-				},
-			},
-			"billing": {
-				Name: "billing",
-				// no Warehouse pin → falls back to cluster default
-			},
-			"dormant": {
-				Name: "dormant",
-				Warehouse: &configstore.ManagedWarehouseConfig{
-					Image: "posthog/duckgres:v1.4.0",
-				},
-				// not in tr.orgs → skipped
-			},
-		},
-	}
-
-	tr.reconcileWarmCapacity(snap)
-
-	got := sharedPool.PerImageWarmTargets()
-	want := map[string]int{
-		"posthog/duckgres:default": 1,
-		"posthog/duckgres:v1.5.1":  1,
-	}
-	if !reflect.DeepEqual(got, want) {
-		t.Fatalf("expected per-image targets %v, got %v", want, got)
-	}
-}
-
-func TestOrgRouterReconcileWarmCapacityTreatsSharedWarmTargetAsDefaultImageBase(t *testing.T) {
-	sharedPool, _ := newTestK8sPool(t, 10)
-
-	tr := &OrgRouter{
-		sharedPool: sharedPool,
-		baseCfg: K8sWorkerPoolConfig{
-			WorkerImage: "posthog/duckgres:default",
-		},
-		globalCfg: ControlPlaneConfig{
-			K8s: K8sConfig{
-				SharedWarmTarget: 4,
-			},
-		},
-		orgs: map[string]*OrgStack{
-			"analytics": {Config: &configstore.OrgConfig{Name: "analytics"}},
-			"billing":   {Config: &configstore.OrgConfig{Name: "billing"}},
-		},
-	}
-
-	snap := &configstore.Snapshot{
-		Orgs: map[string]*configstore.OrgConfig{
-			"analytics": {
-				Name: "analytics",
-				Warehouse: &configstore.ManagedWarehouseConfig{
-					Image: "posthog/duckgres:v1.5.1",
-				},
-			},
-			"billing": {
-				Name: "billing",
-			},
-		},
-	}
-
-	tr.reconcileWarmCapacity(snap)
-
-	got := sharedPool.PerImageWarmTargets()
-	want := map[string]int{
-		"posthog/duckgres:default": 4,
-		"posthog/duckgres:v1.5.1":  4,
-	}
-	if !reflect.DeepEqual(got, want) {
-		t.Fatalf("expected per-image targets %v, got %v", want, got)
-	}
-}
-
-// Warm-pool teardown: SharedWarmTarget=0 means no per-image warm floor, so no
-// neutral workers are pre-spawned (requests foreground-spawn on demand instead).
-func TestOrgRouterReconcileWarmCapacityZeroTargetDisablesWarmPool(t *testing.T) {
-	sharedPool, _ := newTestK8sPool(t, 10)
-	tr := &OrgRouter{
-		sharedPool: sharedPool,
-		baseCfg:    K8sWorkerPoolConfig{WorkerImage: "posthog/duckgres:default"},
-		globalCfg:  ControlPlaneConfig{K8s: K8sConfig{SharedWarmTarget: 0}},
-		orgs: map[string]*OrgStack{
-			"analytics": {Config: &configstore.OrgConfig{Name: "analytics"}},
-		},
-	}
-	snap := &configstore.Snapshot{
-		Orgs: map[string]*configstore.OrgConfig{
-			"analytics": {
-				Name:      "analytics",
-				Warehouse: &configstore.ManagedWarehouseConfig{Image: "posthog/duckgres:v1.5.1"},
-			},
-		},
-	}
-	tr.reconcileWarmCapacity(snap)
-	if got := sharedPool.PerImageWarmTargets(); len(got) != 0 {
-		t.Fatalf("expected no per-image warm targets with SharedWarmTarget=0, got %v", got)
-	}
-}
-
-func TestOrgRouterReconcileWarmCapacitySkipsEmptyClusterDefault(t *testing.T) {
-	sharedPool, _ := newTestK8sPool(t, 10)
-
-	tr := &OrgRouter{
-		sharedPool: sharedPool,
-		baseCfg:    K8sWorkerPoolConfig{}, // WorkerImage unset
-		globalCfg:  ControlPlaneConfig{K8s: K8sConfig{SharedWarmTarget: 1}},
-		orgs: map[string]*OrgStack{
-			"analytics": {Config: &configstore.OrgConfig{Name: "analytics"}},
-		},
-	}
-
-	snap := &configstore.Snapshot{
-		Orgs: map[string]*configstore.OrgConfig{
-			"analytics": {
-				Name: "analytics",
-				Warehouse: &configstore.ManagedWarehouseConfig{
-					Image: "posthog/duckgres:v1.5.1",
-				},
-			},
-		},
-	}
-
-	tr.reconcileWarmCapacity(snap)
-
-	got := sharedPool.PerImageWarmTargets()
-	want := map[string]int{
-		"posthog/duckgres:v1.5.1": 1,
-	}
-	if !reflect.DeepEqual(got, want) {
-		t.Fatalf("expected per-image targets %v, got %v", want, got)
-	}
-}
